server/requirements.txt
Pillow>=9.0.0
pyzmq>=25.0.0
Flask>=2.0.0
flask-cors>=3.0.0